backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
langchain==0.2.5
langchain-community==0.2.5
langchain-google-genai==1.0.6
chromadb==0.5.0
pymupdf==1.24.5
pydantic==2.7.4
python-dotenv==1.0.1
